move CommCareCaseGroup out of case app
into new casegroups app
