+++ corehq/apps/data_interfaces/views.py
@@ -6,13 +6,13 @@ from django.contrib import messages
 from django.core.cache import cache
 from corehq import privileges, toggles
 from corehq.apps.accounting.decorators import requires_privilege_with_fallback
+from corehq.apps.casegroups.models import CommCareCaseGroup
 from corehq.apps.hqwebapp.forms import BulkUploadForm
 from corehq.apps.hqwebapp.templatetags.hq_shared_tags import static
 from corehq.apps.hqwebapp.utils import get_bulk_upload_form
 from dimagi.utils.excel import WorkbookJSONReader, JSONReaderError
 from django.utils.decorators import method_decorator
 from openpyxl.shared.exc import InvalidFileException
-from casexml.apps.case.models import CommCareCaseGroup
 from corehq import CaseReassignmentInterface
 from corehq.apps.data_interfaces.tasks import bulk_upload_cases_to_group, bulk_archive_forms
 from corehq.apps.data_interfaces.forms import (
