+++ corehq/ex-submodules/casexml/apps/case/models.py
@@ -1,3 +1,9 @@
+"""
+Couch models for commcare cases.
+
+For details on casexml check out:
+http://bitbucket.org/javarosa/javarosa/wiki/casexml
+"""
 from __future__ import absolute_import
 from StringIO import StringIO
 import base64
@@ -12,14 +18,14 @@ from django.core.cache import cache
 from django.conf import settings
 from django.core.urlresolvers import reverse
 from django.utils.translation import ugettext as _
-from casexml.apps.case.dbaccessors import get_reverse_indices
-from dimagi.ext.couchdbkit import *
 from couchdbkit.exceptions import ResourceNotFound, ResourceConflict, BadValueError
 from PIL import Image
+
+from casexml.apps.case.dbaccessors import get_reverse_indices
+from dimagi.ext.couchdbkit import *
 from casexml.apps.case.exceptions import MissingServerDate, ReconciliationError
 from corehq.util.couch_helpers import CouchAttachmentsBuilder
 from couchforms.util import is_deprecation, is_override
-from dimagi.utils.chunked import chunked
 from dimagi.utils.django.cached_object import CachedObject, OBJECT_ORIGINAL, OBJECT_SIZE_MAP, CachedImage, IMAGE_SIZE_ORDERING
 from casexml.apps.phone.xml import get_case_element
 from casexml.apps.case.signals import case_post_save
@@ -40,14 +46,6 @@ from dimagi.utils.couch import (
     CouchDocLockableMixIn,
     LooselyEqualDocumentSchema,
 )
-
-
-"""
-Couch models for commcare cases.  
-
-For details on casexml check out:
-http://bitbucket.org/javarosa/javarosa/wiki/casexml
-"""
 
 CASE_STATUS_OPEN = 'open'
 CASE_STATUS_CLOSED = 'closed'
@@ -921,76 +919,6 @@ class CommCareCase(SafeSaveDocument, IndexHoldingMixIn, ComputedDocumentMixin,
         return None
 
 
-import casexml.apps.case.signals
-
-
-class CommCareCaseGroup(Document):
-    """
-        This is a group of CommCareCases. Useful for managing cases in larger projects.
-    """
-    name = StringProperty()
-    domain = StringProperty()
-    cases = ListProperty()
-    timezone = StringProperty()
-
-    def get_time_zone(self):
-        # Necessary for the CommCareCaseGroup to interact with CommConnect, as if using the CommCareMobileContactMixin
-        # However, the entire mixin is not necessary.
-        return self.timezone
-
-    def get_cases(self, limit=None, skip=None):
-        case_ids = self.cases
-        if skip is not None:
-            case_ids = case_ids[skip:]
-        if limit is not None:
-            case_ids = case_ids[:limit]
-        for case_doc in iter_docs(CommCareCase.get_db(), case_ids):
-            # don't let CommCareCase-Deleted get through
-            if case_doc['doc_type'] == 'CommCareCase':
-                yield CommCareCase.wrap(case_doc)
-
-    def get_total_cases(self, clean_list=False):
-        if clean_list:
-            self.clean_cases()
-        return len(self.cases)
-
-    def clean_cases(self):
-        cleaned_list = []
-        for case_doc in iter_docs(CommCareCase.get_db(), self.cases):
-            # don't let CommCareCase-Deleted get through
-            if case_doc['doc_type'] == 'CommCareCase':
-                cleaned_list.append(case_doc['_id'])
-        if len(self.cases) != len(cleaned_list):
-            self.cases = cleaned_list
-            self.save()
-
-    @classmethod
-    def get_by_domain(cls, domain, limit=None, skip=None, include_docs=True):
-        extra_kwargs = {}
-        if limit is not None:
-            extra_kwargs['limit'] = limit
-        if skip is not None:
-            extra_kwargs['skip'] = skip
-        return cls.view(
-            'case/groups_by_domain',
-            startkey=[domain],
-            endkey=[domain, {}],
-            include_docs=include_docs,
-            reduce=False,
-            **extra_kwargs
-        ).all()
-
-    @classmethod
-    def get_total(cls, domain):
-        data = cls.get_db().view(
-            'case/groups_by_domain',
-            startkey=[domain],
-            endkey=[domain, {}],
-            reduce=True
-        ).first()
-        return data['value'] if data else 0
-
-
 def _action_sort_key_function(case):
     def _action_cmp(first_action, second_action):
         # if the forms aren't submitted by the same user, just default to server dates
@@ -1026,3 +954,7 @@ def _type_sort(action_type):
     Consistent ordering for action types
     """
     return const.CASE_ACTIONS.index(action_type)
+
+
+# import signals
+import casexml.apps.case.signals
